+++ FrontEnd/src/components/MiniComponents/StarForLike.module.css
@@ -1,0 +1,4 @@
+.star {
+  color: #ffd800;
+  font-size: 24px;
+}
